test: refundableAmountOf function

+++ test/integration/streamed-amount-of/streamedAmountOf.t.sol
@@ -19,13 +19,13 @@ contract StreamedAmountOf_Integration_Test is Integration_Test {
     }
 
     function test_StreamedAmountOf_LastTimeUpdateInThePresent() external givenNotNull givenNotCanceled {
-        uint256 streamedAmount = openEnded.streamedAmountOf(defaultStreamId);
+        uint128 streamedAmount = openEnded.streamedAmountOf(defaultStreamId);
         assertEq(streamedAmount, 0, "streamed amount");
     }
 
     function test_StreamedAmountOf() external givenNotNull givenNotCanceled {
         vm.warp({ newTimestamp: WARP_ONE_MONTH });
-        uint256 streamedAmount = openEnded.streamedAmountOf(defaultStreamId);
+        uint128 streamedAmount = openEnded.streamedAmountOf(defaultStreamId);
         assertEq(streamedAmount, ONE_MONTH_STREAMED_AMOUNT, "streamed amount");
     }
 }
